security_scanner/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
gunicorn==24.1.1
idna==3.11
packaging==26.0
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.5.5
urllib3==2.6.3
